Test : dropDatabase() develop, local 환경에서만 작동하도록 수정
        if (process.env.NODE_ENV === 'develop' || process.env.NODE_ENV === 'local') {
            await dataSource.dropDatabase();
        }

+++ src/posts/test/post.repository.spec.ts
@@ -43,7 +43,9 @@ describe('PostRepository (E2E)', () => {
     });
 
     beforeEach(async () => {
-        await dataSource.dropDatabase();
+        if (process.env.NODE_ENV === 'develop' || process.env.NODE_ENV === 'local') {
+            await dataSource.dropDatabase();
+        }
         await dataSource.synchronize();
     })
