+++ src/reviews/test/review.e2e.spec.ts
@@ -38,7 +38,9 @@ describe('ReviewController (E2E)', () => {
     })
 
     beforeEach(async () => {
-        await dataSource.dropDatabase();
+        if (process.env.NODE_ENV === 'develop' || process.env.NODE_ENV === 'local') {
+            await dataSource.dropDatabase();
+        }
         await dataSource.synchronize();
     })
